requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
celery==5.5.1
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
crispy-bootstrap5==2024.10
cryptography==44.0.3
daphne==4.1.2
Django==4.2.11
django-crispy-forms==2.3
django-celery-beat==2.5.0
django-celery-results==2.5.1
frozenlist==1.6.0
geoip2==5.1.0
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
joblib==1.5.0
kombu==5.5.2
maxminddb==2.7.0
multidict==6.4.3
numpy==2.2.5
prompt_toolkit==3.0.51
propcache==0.3.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.22.0
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
redis==5.2.1
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
service-identity==24.2.0
setuptools==80.3.1
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.20.0
zope.interface==7.2
